prediction_service/requirements.txt
fastapi
uvicorn
tensorflow
pydantic
Pillow
numpy